LV1 final

+++ src/main/java/com/sparta/memo/service/MemoService.java
@@ -4,10 +4,13 @@ import com.sparta.memo.dto.MemoRequestDto;
 import com.sparta.memo.dto.MemoResponseDto;
 import com.sparta.memo.entity.Memo;
 import com.sparta.memo.repository.MemoRepository;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class MemoService {
@@ -21,10 +24,8 @@ public class MemoService {
     public MemoResponseDto createMemo(MemoRequestDto requestDto) {
         // RequestDto -> Entity
         Memo memo = new Memo(requestDto);
-
         // DB 저장
         Memo saveMemo = memoRepository.save(memo);
-
         // Entity -> ResponseDto
         MemoResponseDto memoResponseDto = new MemoResponseDto(saveMemo);
 
@@ -33,28 +34,38 @@ public class MemoService {
 
     public List<MemoResponseDto> getMemos() {
         // DB 조회
-        return memoRepository.findAll().stream().map(MemoResponseDto::new).toList();
+        return memoRepository.findAllByOrderByModifiedAtDesc().stream().map(MemoResponseDto::new).toList();
+    }
+
+    public Optional<Memo> getMemoById(Long id) {
+        return memoRepository.findById(id);
     }
 
     @Transactional
-    public Long updateMemo(Long id, MemoRequestDto requestDto) {
-        // 해당 메모가 DB에 존재하는지 확인
+    public MemoResponseDto updateMemo(Long id, MemoRequestDto requestDto) {
         Memo memo = findMemo(id);
+        if(checkPassword(memo.getPassword(), requestDto.getPassword())){
+            memo.update(requestDto);
+        }else{
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "비밀번호가 틀렸습니다.");
+        }
 
-        // memo 내용 수정
-        memo.update(requestDto);
-
-        return id;
+        return new MemoResponseDto(memo);
     }
 
-    public Long deleteMemo(Long id) {
-        // 해당 메모가 DB에 존재하는지 확인
+    public String deleteMemo(Long id , MemoRequestDto requestDto) {
         Memo memo = findMemo(id);
+        if(memo.getPassword()==requestDto.getPassword()){
+            memoRepository.delete(memo);
+        }else{
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "비밀번호가 틀렸습니다.");
+        }
+        return "Success";
+    }
 
-        // memo 삭제
-        memoRepository.delete(memo);
 
-        return id;
+    private boolean checkPassword(int a, int b){
+        return a==b;
     }
 
     private Memo findMemo(Long id) {
